refactor lots into util.py
